using FossilExcavationEvent
