+++ src/main/java/at/hannibal2/skyhanni/events/mining/FossilExcavationEvent.kt
@@ -1,0 +1,5 @@
+package at.hannibal2.skyhanni.events.mining
+
+import at.hannibal2.skyhanni.events.LorenzEvent
+
+class FossilExcavationEvent(val loot: List<Pair<String, Int>>): LorenzEvent()
